Refatora comandos em cogs e adiciona novas funcionalidades
Move comandos principais para cogs modulares (combate, habilidades, mestre, players, sistema), melhora organização do projeto e separa lógicas específicas em arquivos dedicados. Adiciona novos comandos e funcionalidades, como sistema de habilidades, eventos, XP automático, e melhorias no fluxo de criação de ficha e combate. Simplifica o main.py, centralizando apenas inicialização e carregamento de cogs.
